+++ backend/sswm/src/main/java/com/ground/sswm/tree/domain/TreeRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TreeRepository  extends JpaRepository<Tree, Integer> {
+public interface TreeRepository  extends JpaRepository<Tree, Long> {
 }
